refactor(web): thin main entrypoint and clean build output
